Migrate onboarding pages to components/pages for SPA
